Feature#164 create life sim web (#165)
* [Feature#164] Update .gitignore to ignore life-sim-web js crap.

* [Feature#164] Create aetherion webclient.

* [Feature#164] Create life-sim-web.

* [Feature#164] Save life-sim-web progress.

* [Feature#164] Save life-sim-web progress.

* [Feature#164] life-sim-web - Fix create character form .

* [Feature#164] life-sim-web - Create subscribe request.

* [Feature#164] life-sim-web - Fix high level deserialization lz4 and msgpack.

* [Feature#164] life-sim-web - Logs and fetch the correct perception from full response.

* [Feature#164] life-sim-web - Checkpoint - just before deserealization.

* [Feature#164] life-sim-web - Create aetherion webclient decoding tests.

* [Feature#164] life-sim-web - save progress aetherion webclient.

* [Feature#164] life-sim-web - save progress aetherion webclient.

* [Feature#164] life-sim-web - save progress aetherion webclient and life-sim-web.

* [Feature#164] life-sim-web - Checkpoint, entities being decoded and working fine - printing something about the state on the screen

* [Feature#164] life-sim-web - Checkpoint, rendering proper assets

* [Feature#164] life-sim-web - Fix sprite scaling

* [Feature#164] life-sim-web - Refactor Application object to be centralized

* [Feature#164] life-sim-web - Refactor and add entity rendering

* [Feature#164] life-sim-web - Add multiple types of terrain

* [Feature#164] life-sim-web - Added correct water rendering

* [Feature#164] life-sim-web - Saving work for input processing

* [Feature#164] life-sim-web - Saving work for input processing

* [Feature#164] life-sim-web - Saving work for input processing

* [Feature#164] life-sim-web - Saving work for refactoring style of the website

* [Feature#164] life-sim-web - Create minimal version of buttons and windows.

* [Feature#164] life-sim-web - Fix dragging of basic window.

* [Feature#164] Aetherion WASM - Remove old emcc bindings.

* [Feature#164] Aetherion WASM - Refactor and add refactor deserialization.

* [Feature#164] Aetherion and lifesim - Refactor and fix memory leaks.

* [Feature#164] lifesim - refactor get_mass to be get_physics_stats.

* [Feature#164] lifesim - Tries to improve client server performance.

* [Feature#164] lifesim - Implement new websocket streams for receiving state from the server.

* [Feature#164] lifesim - Refactor websocket state stream and create class receiver.

* [Feature#164] lifesim - Use stream version on the browser.

* [Feature#164] lifesim - Keep refactoring server client integrations.

* [Feature#164] lifesim - Remove old code.

* [Feature#164] Refactor Camera, add rendering sprite per section and quadrant, make the player at the center of the voxel.

* [Feature#164] Ignore long lines on the ruff linter.

+++ aetherion/src/components/ConsoleLogsComponent.hpp
@@ -11,7 +11,7 @@
 
 struct ConsoleLogsComponent {
     std::map<std::string, std::string> log_buffer;
-    size_t max_size;
+    uint32_t max_size;
 
     // Helper function to get current time as string
     std::string current_time_string() const {
